refactor(web): rename Accessibility organism to PublicAPI, update routes and menu
